requirements.txt
Flask
numpy
Pillow
google-cloud-storage
tensorflow
keras
